Starting from the 2.18.50.0.4 release of binutils, the x86 assembler no longer accepts `fnstsw %eax': `fnstsw %ax' should be used instead.

+++ src/fpu-ia32.inlines.hh
@@ -81,7 +81,7 @@ fpu_set_control(int c) {
 
 inline int
 fpu_get_status() {
-  int sw;
+  unsigned short sw;
   __asm__ __volatile__ ("fnstsw %0" : "=a" (sw) : : "memory");
   return sw;
 }
